Exception Manager Animated Splashscreen Creation of Card App Icon Glide Integration Room Integration Offline mode

+++ data/src/main/java/com/juandelarosa/data/api/responses/ApiResponse.kt
@@ -36,7 +36,7 @@ data class StyleApi(
 )
 
 data class FontApi(
-    val size: Int?
+    val size: Float?
 )
 data class SizeApi(
     val height: Int?,
